requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.13.0
undetected-chromedriver==3.5.4
supabase==2.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
